refactor(websocket-examples): move Main back to template-local components
Main renders language-specific main-function bodies that are not reused
across templates, so per AGENTS.md #4.5 it belongs in each template's
local components/ directory rather than in @asyncapi/generator-components.

- Delete Example.js, Main.js and their tests/snapshots from the shared
  components package; each websocket template (dart, javascript, python)
  now has its own components/Main.js and composes Imports/Handlers/
  OpenConnection/Main/Close directly in its example.*.js entry.
- Drop the now-unused needsTime prop from the shared Imports component.
- Replace {@link X} JSDoc references in Close and OpenConnection with
  backtick code refs so the docs website's MDX parser does not treat
  them as JS expressions inside list items (fixes the unexpected-lazy
  micromark error on api_components.mdx).
- Regenerate api_components.md.
