+++ packages/components/src/components/example/Imports.js
@@ -10,9 +10,10 @@ const stripExtension = (fileName, ext) => fileName.replace(ext, '');
 
 /**
  * Per-language renderer for the top-of-file imports block in the runnable
- * example script. Each entry receives `{ clientName, clientFileName, needsTime }`
- * and returns the source for that language. Dart ignores `clientName` and
- * `needsTime`; Python uses `needsTime` to toggle the `import time` line.
+ * example script. Each entry receives `{ clientName, clientFileName }` and
+ * returns the source for that language. Dart ignores `clientName`; Python
+ * always imports `time` because the generated example's send loop and
+ * long-running receive wait both call `time.sleep(...)`.
  *
  * @type {Record<Language, Function>}
  */
@@ -21,14 +22,10 @@ const importsConfig = {
     const moduleName = stripExtension(clientFileName, '.js');
     return `const ${clientName} = require('./${moduleName}');`;
   },
-  python: ({ clientName, clientFileName, needsTime }) => {
+  python: ({ clientName, clientFileName }) => {
     const moduleName = stripExtension(clientFileName, '.py');
-    const lines = [];
-    if (needsTime) {
-      lines.push('import time');
-    }
-    lines.push(`from ${moduleName} import ${clientName}`);
-    return lines.join('\n');
+    return `import time
+from ${moduleName} import ${clientName}`;
   },
   dart: ({ clientFileName }) => `import 'dart:async';
 import '${clientFileName}';`,
@@ -41,7 +38,6 @@ import '${clientFileName}';`,
  * @param {Language} props.language - Target programming language.
  * @param {string} props.clientFileName - The generated client file name (e.g. `myClient.js`, `my_client.py`, `my_client.dart`).
  * @param {string} [props.clientName] - The generated client class/symbol name. Required for JavaScript/Python; unused for Dart.
- * @param {boolean} [props.needsTime=false] - Python-only: when true, prepends `import time` (used when the example contains a send loop or a long-running receive wait).
  * @returns {JSX.Element} A `Text` component containing the imports block.
  * @throws When the specified language is not supported.
  *
@@ -54,14 +50,13 @@ import '${clientFileName}';`,
  *       language="python"
  *       clientName="EchoClient"
  *       clientFileName="echo_client.py"
- *       needsTime
  *     />
  *   );
  * }
  *
  * renderImports();
  */
-export function Imports({ language, clientName, clientFileName, needsTime = false }) {
+export function Imports({ language, clientName, clientFileName }) {
   const supportedLanguages = Object.keys(importsConfig);
   const buildSource = importsConfig[language];
 
@@ -69,7 +64,7 @@ export function Imports({ language, clientName, clientFileName, needsTime = fals
     throw unsupportedLanguage(language, supportedLanguages);
   }
 
-  const source = buildSource({ clientName, clientFileName, needsTime });
+  const source = buildSource({ clientName, clientFileName });
 
   return (
     <Text newLines={2}>
